+++ aa_theme_slate/constants.py
@@ -1,0 +1,9 @@
+"""
+Constants used in this app
+"""
+
+# Standard Library
+import os
+
+SLATE_BASE_DIR = os.path.join(os.path.dirname(__file__))
+SLATE_STATIC_DIR = os.path.join(SLATE_BASE_DIR, "static", "aa_theme_slate")